히스토리 업데이트
업데이트 :
- 이제 메세지로 확인한 유사도도 히스토리에서 확인 할 수 있습니다
- 받은 초성을 확인할 수 있는 히스토리 UI를 만들었습니다
- 이제 wordeaterinfo에서 받은 관련 단어를 저장합니다
- 히스토리에 각 페이지를 띄우는 방식을 수정하였습니다

버그 픽스 :
- 간헐적으로 히스토리 앞 뒤 페이지가 넘어가지 않던 문제를 해결했습니다
- 받은 관련 단어가 있을 때 히스토리에 제대로 뜨지 않던 문제를 수정했습니다

+++ Assets/02. Scripts/Managers/FileManager.cs
@@ -45,6 +45,7 @@ public class FileManager : MonoBehaviour
         public int Level;
         public string Answer;
         public string History;
+        public string RelevantLine;
         public List<string> Relevant; //관련 단어
         public string Name;
         public string ImgId;
@@ -110,7 +111,7 @@ public class FileManager : MonoBehaviour
     // ========================================================================
     // [Part 1] 워드이터 게임 데이터 (레벨, 정답, 히스토리)
     // ========================================================================
-    public void SaveWordEaterInfo(int le, string an, string hi, List<string> RR , string id)
+    public void SaveWordEaterInfo(int le, string an, string hi, List<string> RR , string id , string RRL)
     {
         WordEaterData data = new WordEaterData
         {
@@ -118,6 +119,7 @@ public class FileManager : MonoBehaviour
             Answer = an,
             History = hi,
             Relevant = RR,
+            RelevantLine = RRL,
             Name = CurrentPlayerName,
             ImgId = id
         };
@@ -151,6 +153,7 @@ public class FileManager : MonoBehaviour
         {
             gamemanager.HistoryLIne = data.History;
             gamemanager.RelevantResult = data.Relevant;
+            gamemanager.RelevantLine = data.RelevantLine;
         }
     }
 
@@ -180,6 +183,16 @@ public class FileManager : MonoBehaviour
         string json = File.ReadAllText(WordEaterPath);
         WordEaterData data = JsonUtility.FromJson<WordEaterData>(json);
         data.Relevant = Rel;
+
+        File.WriteAllText(WordEaterPath, JsonUtility.ToJson(data, true));
+    }
+
+    public void SavaRelevantLine(string RRL) {
+        if (!File.Exists(WordEaterPath)) return;
+
+        string json = File.ReadAllText(WordEaterPath);
+        WordEaterData data = JsonUtility.FromJson<WordEaterData>(json);
+        data.RelevantLine = RRL;
 
         File.WriteAllText(WordEaterPath, JsonUtility.ToJson(data, true));
     }
